kinda works, but lots of things to fix still

+++ utils.h
@@ -1,1 +1,2 @@
-#define ARR_LEN(arr) ((int)sizeof(arr) / (int)sizeof((arr)[0]))
+#define ARR_LEN(arr) (sizeof(arr) / sizeof((arr)[0]))
+
